floow bridge modify，pay product invalid check nil
OS bridge 中数据类型判断 有问题。eg : double 误转位float 导致定位距离计算出错
强制全部转换为double
需要在oc 端 函数参数 也必须定义为double，然后在内部做转换

+++ sdk/sdk.h
@@ -17,7 +17,7 @@
 
 + (void) config:(id) data;
 + (void) init:(int) handle;
-+ (void) login:(int) type;
++ (void) login:(double) type;
 + (void) share:(id) data;
 + (void) pay:(id) data;
 + (void) openBrowser:(NSString*)url;
@@ -40,7 +40,7 @@
 
 //pay init
 + (void) iospay_init:(id)data;
-+ (void) iospay_req:(NSString*)pid :(int)numb :(BOOL)force;
++ (void) iospay_req:(NSString*)pid :(double)numb :(BOOL)force;
 + (void) iospay_stop;
 //-------------------------
 + (BOOL) handle_url:(NSURL*)url;
